ui: Refactors our app-view/%app-view component (#7752)
We noticed that it was about time we dug into the app-view component as
the skin of the UI has chnged quite a bit since this component started
and it has now become slightly more consistent. The %app-view CSS has
not changed much and things needed within it have gradually become more
clearer.

Apart from some refactoring of CSS here, the biggest change is the
addition of a `nav` Slot, which gets placed immediately underneath the
title of the page. This contains any main navigation elements for the
page.